FIX: `log()` volatility only on hardware differences
Even the tiniest of differences caused by the hardware could lead to
failures if the `log()` function was called on those values if they are
close to 0. I added a work around for that because there is not really
anything to prevent this / it doesn't occur due to a fault in Grads.